mid-process replacing urlconnection.h with libcurl
Rather than fix the url.h and urlconneciton.h libraries that were
provided with the assignment as .o files, without the .c source,
I decided to write my own small library that uses libcurl to do
the same things, but in a manner more consistent with how libcurl
works than it would be if I were shoe-horning it into the spec.
defined in the original headers.